Move Log::Any::Adapter::Base methods to ::Core
Log::Any::Adapter::Core now consolidates all the base class
functionality into one place, but we don't let PAUSE index it so it
can't be used directly as a dependency.

Instead, Log::Any::Adapter::Base wraps it and adapters
can specify a dependency on that, which is a dependency
on Log-Any-Adapter, which is what we want.