Make Log::Any use Log::Any::Manager
Log::Any and Log::Any::Adapter now rely on Log::Any::Manager for
all adapter management. This eliminates the bootstrap management
that used to be in Log::Any and some of the convoluted call structure
that routed from Log::Any to Log::Any::Adapter to Log::Any::Manager.

Log::Any::Manager::get_logger is renamed to get_adapter for clarity,
but the old method is kept as an alias for backwards compatibility.

Other classes are modified to cope with this change.

Some utility functions moved around to places that seemed more natural
and required fewer cross-package calls.

In several areas, "make_method" utility calls were replaced with
hand-rolled sub creation to avoid needing a utility class for a
two-line bit of code.