fix backcompat with Unix::Syslog
The Unix::Syslog module allowed specifying syslog options as a bitfield
only, but Sys::Syslog requires they be specified as strings. To smooth
over the transition, we now load the constants from Unix::Syslog if we
are presented with a number, and map the constants to Sys::Syslog
option strings.