warn when given invalid log levels in adapters
Invalid log levels for the File, Stderr, and Stdout adapters now
result in a warning, and the default level of "trace" is used.
Previously, no warning would be issued and no logs would be
generated.

This change also incorporates useful `@CARP_NOT` entries which allow us
to use `carp` to report errors from the perspective of the user of
Log::Any (and not internal to Log::Any). This work might not be
complete, so we might have some stuff yet to do here to maximize our
Carp's usefulness.

Fixes #55