Capture adapter, for redirecting messages to a callback/arrayref
This adapter is much like the 'Test' adapter, but is intended
for end users.  It can log to an arrayref or a coderef, and
has several convenient options that allow capturing the entire
structured form of a message, or just the text.  It supports a
log_level filter as well.