reopen closed syslog with correct settings
A misbehaving library may do a full openlog()/syslog()/closelog() dance
despite it being completely unnecessary. In this case, the next time
a log is passed to Log::Any, syslog() will call openlog() automatically
and lose our ident and facility settings.

So, now we try to detect this case by checking a package variable in
Sys::Syslog that gets unset when closelog() is called. This isn't
a perfect solution, but it will prevent the most naive of bad libraries
from causing too many logs to go missing...

Thanks to Tom Christiansen and Grant Street Group for reporting this
issue and providing the patch.