Fix stack trace scrubbing in Log::Any::Proxy::WithStackTrace
The previous logic was all wrong - it extracted the stack trace from
any exception objects sent as log messages, and scrubbed them, but then
threw away the exception objects entirely and just sent the stack trace
as if it was the log message.  The new logic logic scrubs the stack
traces in-place within the exception objects.